backend/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
google-generativeai  # Added for Gemini
diffusers
torch
transformers
pillow
python-decouple
requests